requirements.txt
werkzeug==0.16.0
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
requests
pytest
simplecache
pytest-cov
numpy
